Change publication cover image for title

+++ app/concepts/publication/publications_operations.rb
@@ -15,28 +15,21 @@ class PublicationsOperations < CollectionOperations
   class << self
     def upload_poster(params, id=nil)
       id ||= params[:id]
-      folder = "posters/#{id.to_s}"
-
-      poster = {}
 
       pdfFile = params[:pdfFile]
-      poster[:pdf] = S3.upload(pdfFile.path, folder) if pdfFile
 
-      coverFile = params[:coverFile]
-      if coverFile
-        image = MiniMagick::Image.open(coverFile.path)
-        image.resize('600x800')
-        poster[:cover] = S3.upload(image.path, folder)
+      if pdfFile
+        folder = "posters/#{id.to_s}"
+        pdf = S3.upload(pdfFile.path, folder)
+
+        if params[:id]
+          old_poster = Publication.find(params[:id])
+
+          S3.delete_object("#{folder}/#{File.basename(old_poster.pdf)}") if pdfFile
+        end
       end
 
-      if params[:id]
-        old_poster = Publication.find(params[:id])
-
-        S3.delete_object("#{folder}/#{File.basename(old_poster.pdf)}") if pdfFile
-        S3.delete_object("#{folder}/#{File.basename(old_poster.cover)}") if coverFile
-      end
-
-      poster
+      pdf
     end
   end
 end
